+++ src/features/club-create/container/FormContainer.tsx
@@ -5,11 +5,11 @@ import {
   CreateClubFormField,
   InputField,
 } from '@/features/club-create/components';
-import ImageField from '@/features/club-create/container/ImageField';
 import RadioButtonGroup from '@/features/club-create/container/RadioButtonGroup';
 import DatePickerContainer from '@/features/club-create/container/DatePickerField';
 import { useBookClubForm } from '@/features/club-create/hooks';
 import PopUp from '@/components/pop-up/PopUp';
+import ImageField from '@/features/club-create/container/ImageField/ImageField';
 
 function FormContainer() {
   const {
